+++ src/mantle/database/console/class-seed-command.php
@@ -28,14 +28,21 @@ class Seed_Command extends Command {
 	/**
 	 * Run Database Seeding
 	 */
-	public function handle(): void {
+	public function handle(): int {
 		if ( ! $this->confirm_to_proceed() ) {
-			return;
+			return static::FAILURE;
 		}
 
 		// Disable cache purging.
 		if ( class_exists( 'WPCOM_VIP_Cache_Manager' ) ) {
 			remove_action_validated( 'shutdown', [ \WPCOM_VIP_Cache_Manager::instance(), 'execute_purges' ] );
+		}
+
+		$class = $this->container->get_namespace() . '\\Database\\Seeds\\Database_Seeder';
+
+		if ( ! class_exists( $class ) ) {
+			$this->error( "Database Seeder class not found: {$class}" );
+			return static::FAILURE;
 		}
 
 		$this->container
@@ -45,5 +52,7 @@ class Seed_Command extends Command {
 			->__invoke();
 
 		$this->success( __( 'Database seeding completed.', 'mantle' ) );
+
+		return static::SUCCESS;
 	}
 }