+++ src/mantle/framework/console/class-config-cache-command.php
@@ -76,7 +76,10 @@ class Config_Cache_Command extends Command {
 	 * Boot a fresh copy of the application configuration.
 	 */
 	protected function get_fresh_configuration(): array {
-		$app = require $this->container->get_bootstrap_path( '/app.php' ); // phpcs:ignore WordPressVIPMinimum.Files.IncludingFile.UsingVariable
+		$bootloader = require $this->container->get_bootstrap_path( '/app.php' ); // phpcs:ignore WordPressVIPMinimum.Files.IncludingFile.UsingVariable
+
+		$app = $bootloader->get_application();
+
 		$app->set_base_path( $this->container->get_base_path() );
 		$app->make( Kernel::class )->bootstrap();