+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Attachment factories now properly use the `Attachment_Factory` class.
 - Ensure that the `delete()` method of the HTTP Client doesn't set a body by default.
 
 ## v1.2.0 - 2024-09-23