+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Define a `thumbnail` relationship on the post model that when defined will return an attachment model.
+
 ### Changed
 
 - Post factories that are passed a term slug to `with_terms()` will create the